requirements.txt
torch
torchvision
gradio
cryptography
numpy
pillow
tqdm
pyyaml
datasets